seperated mutations and object types into new files, cleaning up the schema file.
