+++ server/tp_server/tp/statistics.py
@@ -1,4 +1,4 @@
-from .models import Error, Problem, Assignment, Teacher
+from .models import Assignment
 
 # Req: _id = 100, assignment_id = 50 Give me Stats
 # assignment = Assignment()
